Use fseeko64 when compiling with Mingw
When compiling with Mingw on Windows the size of 'off_t' is 4 and
both  HAVE_FSEEKO and HAVE_FSEEKO64 are defined - causing 'fseek'
to be defined as 'fseeko'.

The issue is that fseeko takes an offset of type 'off_t', which is
smaller than 'rs_long_t' and will overflow when working with large
files (the overflow occurs in 'rs_file_copy_cb').

This commit defines 'fseek' as 'fseeko64', if possible, when
compiling with Mingw.