Fix #176 rs_int_len() for large values.
This reverts this to what it was before, with some minor tidy ups. I believe
that this is pretty much the only way to make it work when rs_long_t is of an
unknown size. It's also pretty hard to write tests that will still pass on
platforms where rs_long_t is less than 64bits.