Use optimized rabinkarp_pow() and parallelizable DOMULT4().
Change from a generic uint32_pow() to rabinkarp_pow() optimized using a lookup
table for getting powers of RABINKARP_MULT.

Use a DOMULT4() macro to unroll rabinkarp_update() in a way that means 4
multiplies could be piplined/parallelized if the hardware can do it. Each
multiply is not dependent on the results of the previous multiply.