Make the default blocksize a multiple of the 128 byte blake2b block size.
Note the md4sum blocksize is 64, so using 128 ensures it is an integer
multiple of the strongsum blocksize for both kinds of strongsum.

This makes calculating both blake2b and md4sum strongsums a tiny bit faster,
since they don't need to add a padded tail block when finalizing the sum.