Remove `LIBRSYNC_EXPORT` from functions in trace.c.
This is the only place we have LIBRSYNC_EXPORT on function definitions in *.c
files and they shouldn't be there, as they break compiling tests on windows.

Note the introduction of netint_test is the first test that links trace.c, so
is the first time we've noticed this failure on Windows.