Add support for delta callback API.
This is in a pretty rough initial state, but it does work and passes all
tests. It's far from optimized, and only changes deltas to the new API. A
bunch of stuff is yet to be deprecated away pending other job types using this
API too.

Add deltagen.[ch] which defines and implements the delta generator callback
API. The default rs_deltagen_t generates deltas identical to the current
release. Add some rs_send_t data writers for files, buffers, and jobs. Also
includes an unused rs_bufout_t data writer that buffers output.

Change job.h to add fields needed by the delta generator API. Also change all
match/miss length fields to `int` to reflect that the lengths are constrained
to less than int-sized lengths.

Change delta.c to use an rs_deltagen_t delta generator, fixing it to correctly
flush and not rely on iterator calls to invoke rs_tube_catchup() between
state-function calls. Also make it never accumulte more than
MAX_DATA_LEN=32768 of hits as well as misses before invoking the the callback
API, since we now keep and pass match data to the API as well.

Change emit.[ch] to emit cmds into buffers, not directly to the tube. We need
this so cmds can be sent to different delta-generator backends, not just to
the job stream via the tube.

In netint.[ch] add rs_put_netint() to write netints to a buffer, not directly
to the tube. This is now used by emit.c.