Update initial check.yml to include a full build matrix.
This is the final result of experiments done in
https://github.com/dbaarda/librsync/pull/1 slightly modified to incorporate
suggestions from rizsotto in https://github.com/librsync/librsync/pull/222.
The original experiments had a messy commit history including many dead-ends
so this cleaner patch was created rather than merge all that mess.

It includes testing on ubuntu, macos, and windows for Release builds with a
variety of cmake options, and a Debug build on ubuntu. The options include
testing with clang, Ninja, static libs, and excluding rdiff. It installs all
required packages using apt-get, brew, or vcpkg to fully build and test rdiff
on all platforms.