Add MAX_DELTA_CMD in job.h for buffer scaling.
This gives us a single point for defining the size used for delta commands and
streaming buffers.

In job.h define MAX_DELTA_CMD to be the maximum size of a single delta command
at 64K.

In delta.c use MAX_DELTA_CMD to define MAX_MISS_LEN and use it to get the
minimum readahead size in rs_getinput().

In whole.c use MAX_DELTA_CMD for defining the buffer sizes used for delta and
patch operations.