Remove use of ssize_t which doesn't exist on Windows.
It turns out ssize_t is a Posix thing that doesn't exist on Windows. I
originally started using it in stream.h so that negative values could be used
to indicate errors when iterating through buffers. We don't use or need that,
so we can just use size_t instead.