Fix #248 by putting `#include "config.h"` in most src/*.c files. (#249)
* Fix #248 by putting `#include "config.h"` in all src/*.c files.

The iwyu tool doesn't handle `config.h` files well, and these includes were
incorrectly removed, which breaks things on some platforms. Add them back to
most `src/*.c` files with `/* IWYU pragma: keep */` to make iwyu ignore them.
We skip `src/hashtable.c` because it is a standalone tool that is platform
independent.

Also add `/* IWYU pragma: keep */` to includes in `src/fileutil.c` that are
needed on some platforms but not others. This means we can remove the special
exemptions to skip this file for the `iwyu` and `iwyu-fix` targets in
`CMakeLists.txt`.

Add some explicit typecasts to `rollsum.[ch]`  and `patch.c` to silence warnings on windows.

Update NEWS.md for `config.h` include fixes and added typecasts.